code/switch-watcher/requirements.txt
configparser
netmiko
pyyaml